+++ src/main/java/com/github/thedeathlycow/thermoo/mixin/common/EntityPredicateMixin.java
@@ -1,4 +1,4 @@
-package com.github.thedeathlycow.thermoo.mixin;
+package com.github.thedeathlycow.thermoo.mixin.common;
 
 import com.github.thedeathlycow.thermoo.api.predicate.TemperaturePredicate;
 import com.github.thedeathlycow.thermoo.api.temperature.TemperatureAware;
@@ -10,6 +10,7 @@ import net.minecraft.server.world.ServerWorld;
 import net.minecraft.util.JsonHelper;
 import net.minecraft.util.math.Vec3d;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
@@ -18,6 +19,7 @@ import org.spongepowered.asm.mixin.injection.callback.LocalCapture;
 @Mixin(EntityPredicate.class)
 public class EntityPredicateMixin {
 
+    @Unique
     private TemperaturePredicate thermoo$temperaturePredicate = TemperaturePredicate.ANY;
 
     @Inject(
