+++ src/main/java/com/github/thedeathlycow/thermoo/mixin/common/LivingEntityEnvironmentTickMixin.java
@@ -1,4 +1,4 @@
-package com.github.thedeathlycow.thermoo.mixin;
+package com.github.thedeathlycow.thermoo.mixin.common;
 
 import com.github.thedeathlycow.thermoo.impl.LivingEntityEnvironmentTickImpl;
 import net.minecraft.entity.LivingEntity;
