+++ src/main/java/com/github/thedeathlycow/thermoo/mixin/common/package-info.java
@@ -2,6 +2,6 @@
  * Mixin package for Thermoo. Code here should not be relied on by API users.
  */
 @ApiStatus.Internal
-package com.github.thedeathlycow.thermoo.mixin;
+package com.github.thedeathlycow.thermoo.mixin.common;
 
 import org.jetbrains.annotations.ApiStatus;
