+++ src/main/resources/thermoo.mixins.json
@@ -1,7 +1,7 @@
 {
     "required": true,
     "minVersion": "0.8",
-    "package": "com.github.thedeathlycow.thermoo.mixin",
+    "package": "com.github.thedeathlycow.thermoo.mixin.common",
     "compatibilityLevel": "JAVA_17",
     "mixins": [
         "EntityPredicateMixin",
